LadderRow 클래스를 만들어서 List<List<String>>을 List<LadderRow>로 수정했다.
사다리의 한 row의 라인 리스트를 가지고 있는 RadderRow 클래스를 만들어서 사다리 라인 전체를 가지는 radder를 List<List<String>> 에서 List<LadderRow>로 수정했다.
main 메소드에서 로직을 실행하지 않게 Controller 클래스를 생성했다.
클래스를 나눔으로써 클래스의 역할이 더 명확해지고 코드도 간단해졌다.
