+++ src/main/java/kr/codesquad/Controller.java
@@ -1,0 +1,21 @@
+package kr.codesquad;
+
+import java.io.IOException;
+
+public class Controller {
+
+    public void run() {
+        Input input = new Input();
+
+        try {
+            String nameData = input.inputName();
+            int height = input.inputHeight();
+            input.closeBr();
+
+            Ladder ladder = new Ladder(nameData, height);
+            ladder.printRadder();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
