+++ src/main/java/kr/codesquad/Input.java
@@ -8,18 +8,13 @@ public class Input {
     private BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
     public String inputName() throws IOException {
-        System.out.println("참여할 사람 이름을 입력하세요. (이름은 쉼표(,)로 구분)");
-        String nameInputData = br.readLine();
-
-        return nameInputData;
+        System.out.println("참여할 사람 이름을 입력하세요.(이름은 쉼표(,)로 구분)");
+        return br.readLine();
     }
 
     public int inputHeight() throws IOException {
-        System.out.println("최대 사다리 높이는??");
-        int height = Integer.parseInt(br.readLine());
-        System.out.println();
-
-        return height;
+        System.out.println("사다리 높이를 입력하세요.");
+        return Integer.parseInt(br.readLine());
     }
 
     public void closeBr() throws IOException {
